+++ src/filter/filters-type.js
@@ -1,0 +1,13 @@
+'use strict';
+
+/**
+ * Названия фильтров.
+ * @enum {string}
+ */
+module.exports = {
+  'ALL': 'reviews-all',
+  'RECENT': 'reviews-recent',
+  'GOOD': 'reviews-good',
+  'BAD': 'reviews-bad',
+  'POPULAR': 'reviews-popular'
+};
